requirements.txt
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
crypto==1.4.1
cryptography==3.4.7
Flask==1.1.2
Flask-Login==0.4.1
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud==0.34.0
google-cloud-bigquery==2.13.1
google-cloud-bigquery-storage==2.3.0
google-cloud-core==1.6.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
gunicorn==19.9.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
libcst==0.3.18
mypy-extensions==0.4.3
Naked==0.1.31
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.1.5
pandas-gbq==0.15.0
proto-plus==1.18.1
protobuf==3.15.6
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.10.1
pydata-google-auth==1.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
shellescape==3.8.1
six==1.15.0
tqdm==4.59.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.0
Werkzeug==1.0.1
google-auth==1.28.0

